+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/ScalingSelectionDialogBuilder.java
@@ -18,12 +18,11 @@
  */
 package is.codion.swing.common.ui.dialog;
 
-import is.codion.swing.common.ui.control.Control;
+import is.codion.common.model.CancelException;
 
 import org.jspecify.annotations.Nullable;
 
 import javax.swing.JComponent;
-import java.util.function.Consumer;
 
 /**
  * Builds a dialog for selecting the scaling.
@@ -37,21 +36,10 @@ public interface ScalingSelectionDialogBuilder {
 	ScalingSelectionDialogBuilder owner(@Nullable JComponent owner);
 
 	/**
-	 * @param initialSelection the initally selected font size ratio, default 100%
-	 * @return this builder
-	 */
-	ScalingSelectionDialogBuilder initialSelection(int initialSelection);
-
-	/**
 	 * Displays a dialog allowing the user the select a scaling multiplier.
-	 * @param scalingSelected called when the OK button is pressed
+	 * @param title the dialog title
+	 * @return the selected scaling
+	 * @throws CancelException in case of cancel
 	 */
-	void selectScaling(Consumer<Integer> scalingSelected);
-
-	/**
-	 * Creates a {@link Control} for selecting the scaling.
-	 * @param scalingSelected called when the OK button is pressed
-	 * @return a Control for displaying a dialog for selecting a scaling
-	 */
-	Control createControl(Consumer<Integer> scalingSelected);
+ 	int show(String title);
 }